pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["sdairs"]
license = "Apache License 2.0"

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.66.0"
SQLAlchemy = "^1.4.21"
sqlalchemy-cockroachdb = "^1.3.4"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["argon2"], version = "^1.7.4"}
python-multipart = "^0.0.5"
psycopg2-binary = "^2.9.1"
uvicorn = {extras = ["standard"], version = "^0.18.3"}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[http]
  [http.middlewares]

    [http.middlewares.api-stripprefix.stripPrefix]
      prefixes = ["/api/v1"]

  [http.routers]

    [http.routers.app-http]
      entryPoints = ["http"]
      service = "app"
      rule = "PathPrefix(`/api/v1`)"
      middlewares = ["api-stripprefix"]